Be forgiving with overlong EntString lumps
Some(*) maps have incorrect sizes recorded for their EntString lumps -
the lump size in the header goes beyong the end of the file.
(Probably some botched edit of a compiled map.)
The contained entity string is otherwise fine & correct, so try to be
tolerant here and fix the lump to go to the end of the file.

(*) Motivated by ztnmap1 which is shipped in "Quake 2: Quad Damage",
available on GoG. So there's a decent chance to run into this issue
without even downloading some maps off the net.

(cherry picked from commit 3b13ff8c07beefbe748d6405738d67431cc85546)